remodel the entire project!
This gem was originally written mostly the win32api specs leveraging
a combination of win32 api's, wmi and com calls.  The native calls
required far too many data type points of management that proved
to be more esoteric than readable code.

I looked at FFI and while it is a significant improvement, I am taking
the easy route to get to the functionality into the gem in the
quickest time frame.

This gem is initially designed with configuration management in mind
versus orchestration.

At the time of this commit the rspec needs a rewrite as it is
a flawed approach (written when i knew nothing about rspec)